imageMorph/requirements.txt
customtkinter
Pillow
tkinterdnd2
